qgs3daxis: Factor out Qgs3DWiredMesh
It will also be used by `QgsChunkBoundsEntity` in the next commit.
