+++ src/3d/qgs3daxis.h
@@ -27,15 +27,8 @@
 #include <Qt3DRender/QPickEvent>
 #include <Qt3DRender/QScreenRayCaster>
 #include <QVector3D>
-#include <QVector2D>
 
 #include <Qt3DRender/QLayer>
-#if QT_VERSION < QT_VERSION_CHECK(6, 0, 0)
-#include <Qt3DRender/QBuffer>
-#else
-#include <Qt3DCore/QBuffer>
-#endif
-#include <Qt3DRender/QGeometryRenderer>
 
 #include <QtWidgets/QMenu>
 #include "qgs3dmapsettings.h"
@@ -185,41 +178,6 @@ class _3D_EXPORT Qgs3DAxis : public QObject
     QCursor mPreviousCursor = Qt::ArrowCursor;
     QMenu *mMenu = nullptr;
 
-};
-
-/**
- * \ingroup 3d
- * \brief Geometry renderer for lines, draws a wired mesh
- *
- * \since QGIS 3.26
- */
-class Qgs3DWiredMesh : public Qt3DRender::QGeometryRenderer
-{
-    Q_OBJECT
-
-  public:
-
-    /**
-     * \brief Default Qgs3DWiredMesh constructor
-     */
-    Qgs3DWiredMesh( Qt3DCore::QNode *parent = nullptr );
-    ~Qgs3DWiredMesh() override;
-
-    /**
-     * \brief add or replace mesh vertices coordinates
-     */
-    void setVertices( const QList<QVector3D> &vertices );
-
-  private:
-#if QT_VERSION < QT_VERSION_CHECK(6, 0, 0)
-    Qt3DRender::QGeometry *mGeom = nullptr;
-    Qt3DRender::QAttribute *mPositionAttribute = nullptr;
-    Qt3DRender::QBuffer *mVertexBuffer = nullptr;
-#else
-    Qt3DCore::QGeometry *mGeom = nullptr;
-    Qt3DCore::QAttribute *mPositionAttribute = nullptr;
-    Qt3DCore::QBuffer *mVertexBuffer = nullptr;
-#endif
 };
 
 #endif // QGS3DAXIS_H
